chore: move indexShardedDAG
