+++ packages/blob-index/src/index.js
@@ -1,2 +1,3 @@
 export * as ShardedDAGIndex from './sharded-dag-index.js'
 export * from './digest-map.js'
+export { indexShardedDAG } from './util.js'
